added file-not-a-file check

+++ src/executor_files.py
@@ -50,7 +50,7 @@ def handle_group_3(relative_paths, folder_base_local, folder_base_remote):
     for relative_path, md5 in relative_paths.items():
         relative_path = relative_path.lstrip("/")
         uncrypted_file_path = folder_base_local + "/" + relative_path
-        uncrypted_file_path_renamed = folder_base_local + "/" + "DELETED_" + relative_path
+        uncrypted_file_path_renamed = folder_base_local + "/" + "DELETED_" + utils.get_current_unix_ts() + "_" + relative_path
         os.rename(uncrypted_file_path, uncrypted_file_path_renamed)
 
 
@@ -73,12 +73,15 @@ def handle_group_2_4(relative_paths, folder_base_local, folder_base_remote):
         relative_path = relative_path.lstrip("/")
         md5_decrypted = md5
         encrypted_file_path = folder_base_remote + "/" + relative_path + ".gpg"
-        decrypted_file_contents = decrypt_routine.decrypt_single_file_inmemory(encrypted_file_path)
-        md5_encrypted = utils.md5_from_bytes(decrypted_file_contents)
-        if md5_encrypted == md5_decrypted:
-            os.remove(encrypted_file_path)
+        if os.path.isfile(encrypted_file_path):
+            decrypted_file_contents = decrypt_routine.decrypt_single_file_inmemory(encrypted_file_path)
+            md5_encrypted = utils.md5_from_bytes(decrypted_file_contents)
+            if md5_encrypted == md5_decrypted:
+                os.remove(encrypted_file_path)
+            else:
+                print("writing conflicted file to local folder")
+                current_ts = utils.get_current_unix_ts()
+                utils.write_bytes_to_file(decrypted_file_contents, folder_base_local + "/" + current_ts + "_" + relative_path)
+                os.rename(folder_base_remote + "/" + relative_path + ".gpg", folder_base_remote + "/" + current_ts + "_" + relative_path + ".gpg")
         else:
-            print("writing conflicted file to local folder")
-            current_ts = utils.get_current_unix_ts()
-            utils.write_bytes_to_file(decrypted_file_contents, folder_base_local + "/" + current_ts + "_" + relative_path)
-            os.rename(folder_base_remote + "/" + relative_path + ".gpg", folder_base_remote + "/" + current_ts + "_" + relative_path + ".gpg")
+            print(encrypted_file_path + " is not a file")
